--- .env
@@ -1,1 +1,0 @@
-GITHUB_PROFILE=adamsuk
