+++ .gitignore
@@ -23,7 +23,8 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-# local env files
+# env files
+.env
 .env.local
 .env.development.local
 .env.test.local
