+++ auth/withAuth.js
@@ -3,7 +3,6 @@ import router from 'next/router';
 import firebase from 'firebase/app';
 import 'firebase/auth';
 import initFirebase from './config';
-import Layout from '../components/Layout';
 
 initFirebase();
 const auth = firebase.auth();
@@ -12,15 +11,13 @@ const withAuth = Component => props => {
   useEffect(() => {
     auth.onAuthStateChanged(authUser => {
       if (!authUser) {
-        router.push('/sandbox/login-portal/signin');
+        router.push('/login-portal/signin');
       }
     });
   }, []);
 
   return (
-    <Layout title="Private">
-      <Component {...props} />
-    </Layout>
+    <Component {...props} />
   )
 };
