+++ components/sandbox/index.js
@@ -1,0 +1,16 @@
+
+import Calculator from "./Calculator";
+import MusicPlayer from "./MediaPlayer"
+
+const sandboxes = [
+  {
+    title: "Calculator",
+    component: Calculator
+  },
+  {
+    title: "Music Player",
+    component: MusicPlayer
+  }
+]
+
+export default sandboxes;
