+++ package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "@headlessui/react": "^1.4.3",
     "axios": "^0.24.0",
+    "firebase": "^8.10.0",
     "gray-matter": "^4.0.3",
+    "js-cookie": "^3.0.1",
     "lodash": "^4.17.21",
     "mathjs": "^10.1.0",
     "next": "^12.0.9",
@@ -20,6 +22,7 @@
     "raw-loader": "^4.0.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-firebaseui": "^5.0.2",
     "react-icons": "^4.3.1",
     "react-markdown": "^7.1.2",
     "react-responsive": "^9.0.0-beta.5",
